[#10] refactor: 댓글 작성자 패키지 구조 변경
