--- src/main/java/com/mallang/comment/domain/credential/WriterCredential.java
@@ -1,4 +1,0 @@
-package com.mallang.comment.domain.credential;
-
-public interface WriterCredential {
-}
