+++ src/main/java/com/mallang/comment/domain/writer/AnonymousWriterCredential.java
@@ -1,4 +1,4 @@
-package com.mallang.comment.domain.credential;
+package com.mallang.comment.domain.writer;
 
 public record AnonymousWriterCredential(
         String password
