+++ src/main/java/com/mallang/comment/domain/writer/WriterCredential.java
@@ -1,0 +1,4 @@
+package com.mallang.comment.domain.writer;
+
+public interface WriterCredential {
+}
